refactor(spring): extract shared Postgres test base and remove redundant unboxing
  Pull the duplicated container setup, Spring context lifecycle, and
  helper methods from the three TransactionalXxxAspectPostgresTest
  classes into a new AbstractPostgresTestBase:

  - Static PostgreSQLContainer<?> POSTGRES singleton, ctx, jdbc,
    truncate(), stopContext(), countRows(), and BaseConfig
    (dataSource, jdbcTemplate, txManager, dmxTransactionalAspect)
    now live in one place
  - Each subclass retains only its service field, a @BeforeAll that
    calls initContext(ServiceConfig.class), the service interface/impl,
    and a slim `@Configuration` ServiceConfig with the single service bean
  - Remove redundant .intValue() from TransactionalResultAspectPostgresTest
    .countRows() — auto-unboxing is sufficient, consistent with the other
    two sibling tests
